+++ test/input/example-with-css/example2.js
@@ -1,0 +1,3 @@
+import './global.css'
+
+console.log('example 2');