+++ test/basic.test.ts
@@ -9,6 +9,7 @@ function buildOptions(pluginOptions = {}, overrideBuildOptions = {}) {
     entryPoints: ['test/input/example.js'],
     outdir: 'test/output',
     plugins: [manifestPlugin(pluginOptions)],
+    bundle: true,
   }
 
   return {...defaultBuildOptions, ...overrideBuildOptions};
@@ -107,7 +108,7 @@ test('it should not throw an error if the short name has a different extension',
   expect(metafileContents()).toMatchObject({'index.js': 'index.js', 'index.ts': 'index.js'});
 });
 
-test.only('it should throw an error if there are conflicting outputs when the shortNames option is used', async () => {
+test('it should throw an error if there are conflicting outputs when the shortNames option is used', async () => {
   expect.assertions(2);
 
   try {
@@ -165,6 +166,19 @@ test('it should put the manifest file in the outdir directory when outbase is sp
   await require('esbuild').build(buildOptions({}, {outbase: 'test', entryPoints: ['test/input/pages/home/index.js', 'test/input/pages/about/index.js']}));
 
   expect(fs.existsSync(OUTPUT_MANIFEST)).toBe(true);
+});
+
+test('it should allow multiple entrypoints with same css', async () => {
+  await require('esbuild').build(buildOptions({hash: false},{
+    entryPoints: ['test/input/example-with-css/example.js', 'test/input/example-with-css/example2.js']
+  }));
+
+  expect(metafileContents()).toMatchObject({
+    "test/input/example-with-css/example.js": "test/output/example.js",
+    "test/input/example-with-css/example.css": "test/output/example2.css",
+    "test/input/example-with-css/example2.js": "test/output/example2.js",
+    "test/input/example-with-css/example2.css": "test/output/example2.css",
+  });
 });
 
 test('it should allow an extensionless input', async () => {