+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- CSS 'sibling' files that are imported from js will be included in the manifest.
 
 ## [v0.2.1] - 2021-05-10
 ### Fixed